common
added some more meat to the ut class. the idea is to create a child process
that will act as the client to my acceptor.
still need to see what exactly I want it to do. for now, it just sends data
to my port, and doesnt care about anything.