common
crisis!!!
i cant get the resulting map from the acceptor to the client code
they run in two distinct address spaces. oy vey !!! what to do
im just gonna write to a file on disk, and call it a day with this
testing framework. i have a lot to do on top of that.