common
layed out the foundation for a pl module which will be used as a stub
simulating the oit pl module.
i plan on using it to assert the pl logic i nneed in the machine saver.
as the pl symbols all come from libc [not under my control], ii had to
find a way to _parameteriize_ them with a var desc'ing whether pl was
indeed needed in each case.
I thought of the following plan:
for each call where pl is needed, a file will be written to disc.
all syms in the stub module will assert the existance of these files.
each file will bare the symbol's name [.e.g. open_file].