common nw
finished with the basic abstraction of the: nw_message;header and
their respective testers. now, inorder to handle a new protocol, all
you have to do, is create its header, and go wild. of course, as the
fw is still under construction - dont go too wild. Acceptor still has
a lot of debugging to undergo