mostly tools
this commit contains a lot of stuff im doing for
dev env. i added the plugin mechanism to my vim, and
finally was able to install YouCompleteMe.
Now we need to find a way to index AOSP for it
generate those py files, and maybe look into more plugins
that could be useful.
there are some minor stuff in the dot-files too.